Parse and execute IDL for decode(); refactor
