+++ tools/python/encode.py
@@ -2,14 +2,10 @@
 # Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
 # SPDX-License-Identifier: BSD-3-Clause-Clear
 import argparse
-import functools
 import re
-import subprocess
 import sys
-import tempfile
-from pathlib import Path
 
-import yaml
+import idl
 
 import spec
 
@@ -447,442 +443,6 @@ def parse_assembly_arguments(line, instruction_operands):
     return operand_values
 
 
-def builtin_encode_fence_scope(scope):
-    scope_map = {
-        "i": 0b1000,
-        "o": 0b0100,
-        "r": 0b0010,
-        "w": 0b0001,
-        "io": 0b1100,
-        "ir": 0b1010,
-        "iw": 0b1001,
-        "or": 0b0110,
-        "ow": 0b0101,
-        "rw": 0b0011,
-        "ior": 0b1110,
-        "iow": 0b1101,
-        "irw": 0b1011,
-        "orw": 0b0111,
-        "iorw": 0b1111,
-    }
-    if scope in scope_map:
-        return scope_map[scope]
-    print(f'# ERROR: invalid fence scope "{scope}"')
-    return None
-
-
-def builtin_encode_rounding_mode(rm):
-    rm_map = {"rne": 0b000, "rtz": 0b001, "rdn": 0b010, "rup": 0b011, "rmm": 0b100, "dyn": 0b111}
-    if rm in rm_map:
-        return rm_map[rm]
-    print(f'# ERROR: invalid rounding mode "{rm}"')
-    return None
-
-
-def builtin_encode_reg_list(ops):
-    if "reg_list" not in ops:
-        return None
-
-    args = ops["reg_list"].strip().replace(" ", "")
-
-    if args in ["{ra}", "{x1}"]:
-        return 4
-    if args in ["{ra,s0}", "{x1,x8}"]:
-        return 5
-    if args in ["{ra,s0-s1}", "{x1,x8-x9}"]:
-        return 6
-    if args in ["{ra,s0-s2}", "{x1,x8-x9,x18}"]:
-        return 7
-    if args in ["{ra,s0-s3}", "{x1,x8-x9,x18-x19}"]:
-        return 8
-    if args in ["{ra,s0-s4}", "{x1,x8-x9,x18-x20}"]:
-        return 9
-    if args in ["{ra,s0-s5}", "{x1,x8-x9,x18-x21}"]:
-        return 10
-    if args in ["{ra,s0-s6}", "{x1,x8-x9,x18-x22}"]:
-        return 11
-    if args in ["{ra,s0-s7}", "{x1,x8-x9,x18-x23}"]:
-        return 12
-    if args in ["{ra,s0-s8}", "{x1,x8-x9,x18-x24}"]:
-        return 13
-    if args in ["{ra,s0-s9}", "{x1,x8-x9,x18-x25}"]:
-        return 14
-    if args in ["{ra,s0-s11}", "{x1,x8-x9,x18-x27}"]:
-        return 15
-
-
-def builtin_encode_stack_adj(ops, xlen):
-    reg_list = builtin_encode_reg_list(ops)
-    if reg_list is None:
-        return None
-    registers = 13 if reg_list == 15 else (reg_list - 3)
-    registers_space = registers * int(xlen / 8)
-    registers_space_aligned = int((registers_space + 15) / 16) * 16
-    extra_space = (-ops["stack_adj"]) - registers_space_aligned
-    return extra_space >> 4
-
-
-def builtin_encode_sreg(reg):
-    return (reg - 8) - 8 * ((reg - 8) // 10)
-
-
-def builtin_encode_float_immediate(value):
-    if value == "-1.0":
-        return 0b00000
-    if value == "min":
-        return 0b00001
-    if value == "0.0000152587890625":
-        return 0b00010
-    if value == "0.000030517578125":
-        return 0b00011
-    if value == "0.00390625":
-        return 0b00100
-    if value == "0.0078125":
-        return 0b00101
-    if value == "0.0625":
-        return 0b00110
-    if value == "0.125":
-        return 0b00111
-    if value == "0.25":
-        return 0b01000
-    if value == "0.3125":
-        return 0b01001
-    if value == "0.375":
-        return 0b01010
-    if value == "0.4375":
-        return 0b01011
-    if value == "0.5":
-        return 0b01100
-    if value == "0.625":
-        return 0b01101
-    if value == "0.75":
-        return 0b01110
-    if value == "0.875":
-        return 0b01111
-    if value == "1.0":
-        return 0b10000
-    if value == "1.25":
-        return 0b10001
-    if value == "1.5":
-        return 0b10010
-    if value == "1.75":
-        return 0b10011
-    if value == "2.0":
-        return 0b10100
-    if value == "2.5":
-        return 0b10101
-    if value == "3":
-        return 0b10110
-    if value == "4":
-        return 0b10111
-    if value == "8":
-        return 0b11000
-    if value == "16":
-        return 0b11001
-    if value == "128":
-        return 0b11010
-    if value == "256":
-        return 0b11011
-    if value == "32768":
-        return 0b11100
-    if value == "65536":
-        return 0b11101
-    if value == "inf":
-        return 0b11110
-    if value == "nan":
-        return 0b11111
-    print(f'# ERROR: Unrecognized floating point immediate: "{value}"')
-    return None
-
-
-class IdlExecutionError(Exception):
-    pass
-
-
-class _IdlReturn(Exception):
-    def __init__(self, value):
-        self.value = value
-
-
-_REPO_ROOT = Path(__file__).resolve().parents[2]
-_IDLC = _REPO_ROOT / "bin" / "idlc"
-
-
-@functools.cache
-def _compile_idl_function_body(idl):
-    with tempfile.NamedTemporaryFile(
-        "w", encoding="utf-8", suffix=".idl", delete_on_close=False
-    ) as idl_file:
-        idl_file.write(idl)
-        idl_path = idl_file.name
-        idl_file.close()
-
-        result = subprocess.run(
-            [
-                str(_IDLC),
-                "compile",
-                "--format",
-                "yaml",
-                "--root",
-                "function_body",
-                idl_path,
-            ],
-            cwd=_REPO_ROOT,
-            text=True,
-            capture_output=True,
-            check=False,
-        )
-
-    if result.returncode != 0:
-        details = result.stderr.strip() or result.stdout.strip()
-        raise IdlExecutionError(f"IDL compile failed: {details}")
-
-    return yaml.safe_load(result.stdout)
-
-
-def _trunc_div(lhs, rhs):
-    if rhs == 0:
-        raise IdlExecutionError("division by zero")
-    sign = -1 if (lhs < 0) ^ (rhs < 0) else 1
-    return sign * (abs(lhs) // abs(rhs))
-
-
-def _operand_key(node, env):
-    if isinstance(node, dict) and node.get("kind") == "id":
-        return node["name"]
-    return _eval_idl_expr(node, env)
-
-
-def _operand_offset_value(operand_name, env):
-    for operand_def in env.get("operand_defs") or []:
-        if operand_def.get("name") != operand_name:
-            continue
-        offset_def = operand_def.get("offset")
-        if isinstance(offset_def, dict):
-            offset_name = offset_def.get("name")
-            if offset_name in env["operands"]:
-                return env["operands"][offset_name]
-
-    raise IdlExecutionError(f"no offset value found for operand '{operand_name}'")
-
-
-def _format_idl_message(message, env):
-    def replace_operand(match):
-        key = match.group(1)
-        return str(env["operands"].get(key, "??"))
-
-    return re.sub(r"\$\{operands\[([A-Za-z_][A-Za-z0-9_]*)\]\}", replace_operand, message)
-
-
-def _eval_idl_funcall(node, env):
-    func = node["func"]
-    args = [_eval_idl_expr(arg, env) for arg in node.get("args", [])]
-
-    if func == "reg2creg":
-        if len(args) != 1:
-            raise IdlExecutionError("reg2creg expects one argument")
-        return args[0] - 8
-    if func == "creg2reg":
-        if len(args) != 1:
-            raise IdlExecutionError("creg2reg expects one argument")
-        return args[0] + 8
-    if func == "xlen":
-        if args:
-            raise IdlExecutionError("xlen expects no arguments")
-        return env["xlen"]
-    if func == "raise":
-        message = args[0] if args else "IDL raised"
-        if isinstance(message, str):
-            message = _format_idl_message(message, env)
-        raise IdlExecutionError(str(message))
-
-    raise IdlExecutionError(f"unsupported IDL function '{func}'")
-
-
-def _eval_idl_binary(node, env):
-    op = node["op"]
-
-    if op == "||":
-        return bool(_eval_idl_expr(node["lhs"], env)) or bool(_eval_idl_expr(node["rhs"], env))
-    if op == "&&":
-        return bool(_eval_idl_expr(node["lhs"], env)) and bool(_eval_idl_expr(node["rhs"], env))
-
-    lhs = _eval_idl_expr(node["lhs"], env)
-    rhs = _eval_idl_expr(node["rhs"], env)
-
-    if op == "==":
-        return lhs == rhs
-    if op == "!=":
-        return lhs != rhs
-    if op == "<":
-        return lhs < rhs
-    if op == "<=":
-        return lhs <= rhs
-    if op == ">":
-        return lhs > rhs
-    if op == ">=":
-        return lhs >= rhs
-    if op == "+":
-        return lhs + rhs
-    if op == "-":
-        return lhs - rhs
-    if op == "*":
-        return lhs * rhs
-    if op == "/":
-        return _trunc_div(lhs, rhs)
-    if op == "%":
-        return lhs % rhs
-    if op == "<<":
-        return lhs << rhs
-    if op == ">>":
-        return lhs >> rhs
-    if op == "&":
-        return lhs & rhs
-    if op == "|":
-        return lhs | rhs
-    if op == "^":
-        return lhs ^ rhs
-
-    raise IdlExecutionError(f"unsupported IDL binary operator '{op}'")
-
-
-def _eval_idl_expr(node, env):
-    if not isinstance(node, dict):
-        return node
-
-    kind = node.get("kind")
-
-    if kind == "bits_literal":
-        return int(node["value"])
-    if kind == "string_literal":
-        return node["text"]
-    if kind == "id":
-        name = node["name"]
-        if name in env["vars"]:
-            return env["vars"][name]
-        raise IdlExecutionError(f"unknown IDL identifier '{name}'")
-    if kind == "paren_expr":
-        return _eval_idl_expr(node["expr"], env)
-    if kind == "array_access":
-        array = node["array"]
-        if (
-            isinstance(array, dict)
-            and array.get("kind") == "id"
-            and array.get("name") == "operands"
-        ):
-            return env["operands"].get(_operand_key(node["index"], env), "??")
-        array_value = _eval_idl_expr(array, env)
-        return array_value[_eval_idl_expr(node["index"], env)]
-    if kind == "operand_offset_access":
-        return _operand_offset_value(node["operand_name"], env)
-    if kind == "funcall_expr":
-        return _eval_idl_funcall(node, env)
-    if kind == "binary_operator_expr":
-        return _eval_idl_binary(node, env)
-    if kind == "unary_operator_expr":
-        op = node["op"]
-        value = _eval_idl_expr(node["expr"], env)
-        if op == "-":
-            return -value
-        if op == "+":
-            return value
-        if op == "!":
-            return not bool(value)
-        if op == "~":
-            return ~value
-        raise IdlExecutionError(f"unsupported IDL unary operator '{op}'")
-    if kind == "ternary_operator_expr":
-        branch = "true_expression" if _eval_idl_expr(node["condition"], env) else "false_expression"
-        return _eval_idl_expr(node[branch], env)
-    if kind == "var_decl_init":
-        value = _eval_idl_expr(node["value"], env)
-        env["vars"][node["name"]["name"]] = value
-        return value
-    if kind == "var_assignment":
-        value = _eval_idl_expr(node["value"], env)
-        var = node["var"]
-        if var.get("kind") != "id":
-            raise IdlExecutionError("unsupported assignment target")
-        env["vars"][var["name"]] = value
-        return value
-    if kind == "return_expr":
-        exprs = node.get("exprs", [])
-        if len(exprs) == 0:
-            raise _IdlReturn(None)
-        if len(exprs) == 1:
-            raise _IdlReturn(_eval_idl_expr(exprs[0], env))
-        raise _IdlReturn(tuple(_eval_idl_expr(expr, env) for expr in exprs))
-
-    raise IdlExecutionError(f"unsupported IDL expression kind '{kind}'")
-
-
-def _exec_idl_body(node, env):
-    if not isinstance(node, dict):
-        raise IdlExecutionError("invalid IDL body")
-
-    kind = node.get("kind")
-    if kind not in ("function_body", "if_body"):
-        raise IdlExecutionError(f"unsupported IDL body kind '{kind}'")
-
-    for stmt in node.get("stmts", []):
-        _exec_idl_stmt(stmt, env)
-
-
-def _exec_idl_if(node, env):
-    if _eval_idl_expr(node["condition"], env):
-        _exec_idl_body(node["taken_body"], env)
-        return
-
-    for else_if in node.get("else_ifs", []) or []:
-        if _eval_idl_expr(else_if["condition"], env):
-            _exec_idl_body(else_if["body"], env)
-            return
-
-    else_body = node.get("else")
-    if else_body:
-        _exec_idl_body(else_body, env)
-
-
-def _exec_idl_stmt(node, env):
-    if not isinstance(node, dict):
-        raise IdlExecutionError("invalid IDL statement")
-
-    kind = node.get("kind")
-    if kind == "stmt":
-        _eval_idl_expr(node["expr"], env)
-        return
-    if kind == "if_stmt":
-        _exec_idl_if(node, env)
-        return
-
-    raise IdlExecutionError(f"unsupported IDL statement kind '{kind}'")
-
-
-def execute_idl_encode(idl, operands, xlen, instruction_operands=None):
-    ast = _compile_idl_function_body(idl)
-    env = {
-        "operands": operands,
-        "operand_defs": instruction_operands or [],
-        "vars": {},
-        "xlen": xlen,
-    }
-
-    try:
-        _exec_idl_body(ast, env)
-    except _IdlReturn as ret:
-        return ret.value
-
-    return None
-
-
-def UDB_invoke_IDL(idl, operands, xlen, instruction_operands=None):
-    try:
-        return execute_idl_encode(idl, operands, xlen, instruction_operands)
-    except IdlExecutionError as exc:
-        print(f"#  ERROR: IDL encode failed: {exc}")
-        return None
-
-
 def fill_in_variables(inst, assembly, xlen=64):
     """Fill in variables in the match pattern based on the instruction's operands"""
     # Initialize variables
@@ -946,14 +506,15 @@ def fill_in_variables(inst, assembly, xlen=64):
         operand_value = 0  # Default value
 
         if "encode(operands)" in variable:
-            operand_value = UDB_invoke_IDL(
-                variable["encode(operands)"],
-                assembly_operands,
-                xlen,
-                instruction_operands,
-            )
-            if operand_value is None:
-                print("#  ERROR: unsupported variable encoding")
+            try:
+                operand_value = idl.execute(
+                    variable["encode(operands)"],
+                    xlen,
+                    operands=assembly_operands,
+                    instruction_operands=instruction_operands,
+                )
+            except idl.IdlExecutionError as exc:
+                print(f"#  ERROR: IDL encode failed: {exc}")
                 return None
         elif var_name in assembly_operands:
             dprint(
